Crate diesel_util

Re-exports

pub use scoped_futures;

Macros

db_filter

diesel_util_lazy_static

dynamic_schema

Structs

Column

DbConnection

DynamicSchema

IsDeletedFilter

IsNotDeletedFilter

Joinable

JoinablePair

MAX_PAGINATE_COUNT

Page

Paged

Paginated

Table

TxCleanupError

WithTitle

Wrapper

Enums

ColumnSqlType

DbEntityError

wrapper error type returned from DbEntity methods solely exists to facilitate error conversion when converting between the raw db representation and the entity representation

Traits

Audit

CanPaginate

DbDelete

DbEntity

DbGet

DbInsert

DbUpdate

Deletable

Enum

IncludesChanges

IsDeleted

IsNotDeleted

MaybeAudit

PageRef

Paginate

Partition

SchemaExpressionMethods

SoftDeletable

TxFn

_Db

_Db represents a shared reference to a mutable async db connection It abstracts over the case where the connection is owned vs. a mutable reference. The main goal of this abstraction is to defer locking the access to the connection until the latest point possible, allowing other code paths (excepting for connections in transactions) to access the connection as well. At the moment, Db is passed in by value instead of reference into the transaction provided transaction wrapper so you’ll need to use &conn instead of just conn.

Type Definitions

DbConnRef

InsertStmt

TxCleanup

Attribute Macros

soft_delete

container level attribute used by SoftDelete

Derive Macros

Audit

Db

Enum

IncludesChanges

SoftDelete